refactor(test-support): move semantic builders out of runtime
